Separate dapp request errors and WalletConnect responses
