+++ src/features/dapp-request/utils/mobileWalletProtocolErrors.ts
@@ -1,0 +1,6 @@
+export enum MobileWalletProtocolUserErrors {
+  USER_REJECTED_HANDSHAKE = 'User rejected the handshake',
+  USER_REJECTED_REQUEST = 'User rejected request',
+  ABORTED = 'Aborted',
+  READ_ONLY_WALLET = 'This wallet is read-only',
+}
